move format_results function to training utils
